add working unit tests to operator
